requirements.txt
python-telegram-bot==20.7
better_profanity
nest_asyncio

